requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
djangorestframework==3.15.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
psycopg2-binary==2.9.9
pycountry==23.12.11
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
